ENG-1232: add support for enum16 and add ch enum validation (#2972)
<!-- CURSOR_SUMMARY -->
> [!NOTE]
> Adds full signed ClickHouse Enum8/Enum16 support (including negative
values) across parsing, codegen, and validation, with comprehensive
tests and documentation.
> 
> - **Core Types**:
> - Change enum integer representation from `u8` to `i16` in
`EnumValue`/metadata and TS generator (`TSEnumValue.Number(i16)`).
> - Safe proto conversion for enum ints with range checks; add unit
tests for valid/out-of-range cases.
> - **ClickHouse Parsing/Conversion**:
> - Tokenizer now supports signed `NumberLiteral`; Enum regex supports
negatives in `infrastructure/olap/clickhouse/model.rs`.
> - Validate Enum8/Enum16 ranges and convert to `Enum<Int(i16)>` in
`type_parser.rs`.
> - Reject negative JSON params for
`max_dynamic_types`/`max_dynamic_paths`.
> - Add extensive tests for Enum8/16 boundaries, negatives, out-of-range
errors, and JSON param validation.
> - **Python Parser**:
> - Parse integer enum members as `i16` with clear out-of-range errors.
> - **Validation Runtime (`validate_passthrough.rs`)**:
> - Adjust enum integer conversions (`from_i16`) and prevent negative
enum values from matching unsigned (`u64`) inputs.
> - Add tests covering negative enum handling for `u64`/`i64` and map
key validation.
> - **E2E Tests (db-pull)**:
> - Add Enum16 column with large values; verify generated Python/TS code
preserves enum values and default syntax.
> - **Docs**:
> - Document integer-backed enums with examples and Enum8/Enum16 ranges
in `supported-types.mdx`.
> 
> <sup>Written by [Cursor
Bugbot](https://cursor.com/dashboard?tab=bugbot) for commit
ccef56b263bbc9482f1e690ed2e1192eae9c64df. This will update automatically
on new commits. Configure
[here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

+++ apps/framework-cli/src/framework/typescript/generator.rs
@@ -133,7 +133,8 @@ pub struct TSEnumMember {
 #[derive(Debug, Clone, Serialize, Eq, PartialEq, Hash)]
 pub enum TSEnumValue {
     String(String),
-    Number(u8),
+    /// Number value for numeric enums (supports Enum8: -128 to 127, Enum16: -32768 to 32767)
+    Number(i16),
 }
 
 impl fmt::Display for InterfaceFieldType {
